+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 3.5.0 (Unreleased)
 
 ### Added
-- Support for (Cross Region Copy)[https://docs.cloud.oracle.com/iaas/Content/Object/Tasks/copyingobjects.htm] of objects
+- Support for [Cross Region Copy](https://docs.cloud.oracle.com/iaas/Content/Object/Tasks/copyingobjects.htm) of objects
+- Support for object lifecycle policies on a bucket on object storage. See [Using Object Lifecycle Management](https://docs.cloud.oracle.com/iaas/Content/Object/Tasks/usinglifecyclepolicies.htm)
 
 ## 3.4.0 (October 11, 2018)